+++ changelog.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - segfault because of `input_id` narrowing conversion
 - unhandled exception when ConfFile doesn't exist
 - don't try to render reports that failed to load
+- loader logofile path can be something other than `/usr/viewtouch/graphics/logofile`
+- loader CSS deprecation warnings
 
 
 ## [v19.02.1] - 2019-02-16